Fixing SendInput and its related types (suggestions on AArnott#245)
