+++ src/User32.Desktop/User32+INPUT.cs
@@ -25,19 +25,5 @@ namespace PInvoke
                 get { return Marshal.SizeOf(typeof(INPUT)); }
             }
         }
-
-        /// <summary>
-        /// Union meant to help correctly marshalling <see cref="INPUT" /> with the different input types.
-        /// </summary>
-        [StructLayout(LayoutKind.Explicit)]
-        public struct InputUnion
-        {
-            [FieldOffset(0)]
-            public MOUSEINPUT mi;
-            [FieldOffset(0)]
-            public KEYBDINPUT ki;
-            [FieldOffset(0)]
-            public HARDWAREINPUT hi;
-        }
     }
 }
