+++ src/User32.Desktop/User32+InputUnion.cs
@@ -1,0 +1,27 @@
+﻿// Copyright (c) to owners found in https://github.com/AArnott/pinvoke/blob/master/COPYRIGHT.md. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
+
+namespace PInvoke
+{
+    using System.Runtime.InteropServices;
+
+    /// <content>
+    /// Contains the <see cref="InputUnion"/> nested type.
+    /// </content>
+    public partial class User32
+    {
+        /// <summary>
+        /// Union meant to help correctly marshalling <see cref="INPUT" /> with the different input types.
+        /// </summary>
+        [StructLayout(LayoutKind.Explicit)]
+        public struct InputUnion
+        {
+            [FieldOffset(0)]
+            public MOUSEINPUT mi;
+            [FieldOffset(0)]
+            public KEYBDINPUT ki;
+            [FieldOffset(0)]
+            public HARDWAREINPUT hi;
+        }
+    }
+}
